% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{checkApprovalStatus}
\alias{checkApprovalStatus}
\title{checkApprovalStatus(d, d2_session)}
\usage{
checkApprovalStatus(d, d2_session)
}
\arguments{
\item{d2_session}{the d2_session.}

\item{r}{the response from a post to DATIM.}
}
\value{
a summary of the import process.
}
\description{
Returns a summary of the import process for a user to check if there were any issues.
}
