% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getBaseURL}
\alias{getBaseURL}
\title{Title getBaseURL()}
\usage{
getBaseURL()
}
\value{
A base URL to be used throughout the application. If the BASE_URL
environment variable is not set,
this function will return https://www.datim.org,
otherwise, the value of the environment variable.
}
\description{
Title getBaseURL()
}
