% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getCountryNameFromUID}
\alias{getCountryNameFromUID}
\title{getCountryNameFromUID(country_uids)}
\usage{
getCountryNameFromUID(uid)
}
\arguments{
\item{r}{the response from a post to DATIM.}

\item{d2_session}{the d2_session.}
}
\value{
a summary of the import process.
}
\description{
Returns a summary of the import process for a user to check if there were any issues.
}
