% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importToDatim.R
\name{importToDatim}
\alias{importToDatim}
\title{import(server, deletes_json, main_import_json, dedupes_0000_json, dedupes_00001_json)}
\usage{
importToDatim(d, server, import_data_json, import_data, d2session)
}
\arguments{
\item{d}{the datapack object which holds all relevant datapack information and validation results.}

\item{server}{the server we are importing to like triage or production.}

\item{import_data_json}{the import data object which holds the 4 types of import data in its prepped json format.}

\item{import_data}{the import data object which holds the 4 types of import data.}
}
\value{
printed information.
}
\description{
runs import procedure for all data into DATIM.
}
