% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollImportStatus.R
\name{pollImportStatus}
\alias{pollImportStatus}
\title{pollImportStatus(r)}
\usage{
pollImportStatus(r)
}
\arguments{
\item{r}{The response from a post to DATIM.}
}
\value{
printed statements on data
}
\description{
Returns information on the status of an import.
}
