% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_json.R
\name{prep_json}
\alias{prep_json}
\title{prep_json(data)}
\usage{
prep_json(data)
}
\arguments{
\item{data}{an import file.}
}
\value{
data
}
\description{
Preps a json file based off the import files as data inputs. These are the jsons that are imported.
}
