% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepImportFiles.R
\name{prepImportFiles}
\alias{prepImportFiles}
\title{prepImportFiles(data, tool, type)}
\usage{
prepImportFiles(d, tool, type)
}
\arguments{
\item{d}{a d object result typically from a datapack being unpacked.}

\item{tool}{the type of tool, Data Pack or OPU Data Pack}

\item{the}{type of}
}
\value{
data
}
\description{
Preps the 4 import files
}
