% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepJson.R
\name{prepJson}
\alias{prepJson}
\title{prepJson(data)}
\usage{
prepJson(data)
}
\arguments{
\item{data}{an import file.}
}
\value{
data
}
\description{
Preps a json file based off the import files as data inputs. These are the jsons that are imported.
}
