% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendPayload.R
\name{sendPayload}
\alias{sendPayload}
\title{sendPayload(data, type, d2_session)}
\usage{
sendPayload(data, type, d2_session)
}
\arguments{
\item{data}{an import file.}

\item{type}{an import file.}

\item{d2_session}{a d2 session}
}
\value{
data
}
\description{
....
}
