% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{validate(file, d2_session)}
\usage{
validate(file, d2_session)
}
\arguments{
\item{file}{The path of the datapack to parse.}

\item{d2_session}{The d2 session object.}
}
\value{
d object holding all datapack information.
}
\description{
Parses a datapack and returns datpack information as well as import file data.
}
