% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importToDatim.R
\name{importToDatim}
\alias{importToDatim}
\title{import(server, deletes_json, main_import_json, dedupes_0000_json, dedupes_00001_json)}
\usage{
importToDatim(d, server, import_data, d2session)
}
\arguments{
\item{server}{the server we are importing to like triage or production.}

\item{deletes_json}{the deletes json file used to delete existing data before import.}

\item{main_import_json}{the main imports imported to datim.}

\item{dedupes_00000_json}{the pure dedupe data 0000 imported to datim.}

\item{dedupes_00001_json}{the pure dedupe data 0001 imported to datim.}
}
\value{
data
}
\description{
runs import procedure for all data into DATIM.
}
