% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepImportFiles.R
\name{prepImportFiles}
\alias{prepImportFiles}
\title{prepImportFiles(data)}
\usage{
prepImportFiles(d, tool, type)
}
\arguments{
\item{data}{an import file.}
}
\value{
data
}
\description{
Preps a json file based off the import files as data inputs. These are the jsons that are imported.
}
